Create getComputerChoice function
This function randomly returns the strings "rock",
"paper" or "scissors".

This is achieved by creating the variable
computerChoice.
The variable uses the Math.random() to create
a random number between 0 and 0,99.
Then the generated number is multiplied by 3 to
increase the random number range (0 to 2,97).
Finally, Math.floor() rounds down the numbers
returning the largest integer (0, 1, 2) providing
3 choices.

The if..else construction tests the value of the
variable computerChoice and returns the strings
"rock", "paper" and "scissors" when the variable
is either 0, 1 or 2 respectively.