Create playGame() function
Move computerScore and humanScore variables inside the playGame()
function.

Move playRound() function inside playGame() function.

Declared the variable countRound within the scope of playGame()
function.

Add increments of countRound for each if..else whithin the playRound()
function.

Add a function for each round of the game (5 rounds). Each of this
functions store the choices of the computer and the human in the
variables computerSelection and humanSelection. Then call playRound().
This seems something to refactor when I learn about loops.