Add if..else statement to check game winner
Add if..else statement within the playRound() function scope
and after the 5 function calls for each round. This way, the if..else
statement is only executed after the 5 rounds have completed.

It checks humanScore vs. computerScore to alert the user whether
there's a winner (computer or human) or if the game ends in a draw.