Add user input check
Write an if..else statement within the function getHumanChoice()
that checks whether the user's input is correct ("rock", "paper",
"scissors"). If the input is not valid, then alert user and run the
function again until the input is valid.

Before this change any other string would be treated as a valid input.
As a consequence, the function playRound() would accept any string and
would always count this round in favor of the Human (as it would always
treat the condition as scissors vs. paper, see last else of the
if..else statement wihtin playRound()).